+++ my-app/src/components/Footer/Footer.style.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { Layout } from 'antd';
-import { Main_color } from '../../const/const'
+import { Main_color } from '../../const/actionConst'
 const { Footer } = Layout;
 
 export const FooterWithStyles = styled(Footer) `
